Keep factory impl in is own class Delegate in RTCVideoRenderer
